requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
channels==4.1.0
channels-redis==4.2.0
Django==5.0.6
djangorestframework==3.15.1
msgpack==1.0.8
redis==5.0.4
sqlparse==0.5.0
typing_extensions==4.11.0
